pyproject.toml
[build-system]
requires = ["setuptools==69.5.1", "wheel==0.43.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-quant-system"
version = "0.1.0"
description = "Crypto quant system scaffold with optional addon modules."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Repo Builder"}]
dependencies = [
  "numpy==1.26.4",
  "pandas==2.2.2",
  "pyyaml==6.0.1",
  "scikit-learn==1.4.2",
  "fastapi==0.111.0",
  "uvicorn==0.30.1",
]

[project.optional-dependencies]
test = ["pytest==8.2.2"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
